requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.0.0
torch>=2.0.0
openai-whisper>=1.0.0
pydub>=0.25.1
moviepy>=1.0.3
yt-dlp>=2023.0.0
requests>=2.31.0
stripe>=7.0.0
python-jose>=3.3.0
passlib>=1.7.4
fastapi>=0.100.0
bcrypt>=4.0.1
python-multipart>=0.0.6
uvicorn>=0.23.0
psycopg2-binary>=2.9.9 